Introduce FixtureSet::ModelMetadata
